fix state issues with having extra dependency
